admin-back/users/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from users.views import UserView

router = DefaultRouter()
router.register('users', UserView)

urlpatterns = [
    path('', include(router.urls)),
]