admin-back/requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
coreschema==0.0.4
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mysqlclient==2.2.8
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
setuptools==82.0.1
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
